debundle/lowering: extract util
BAZEL_TEST_INVOCATIONS=none: file move.
